requirements.txt
streamlit
playwright==1.38.0
SQLAlchemy==2.0.19
psycopg2-binary==2.9.7
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
google-cloud-translate
langdetect
pandas
bertopic
sklearn
google
fastapi
uvicorn
logging